feat: add testing configuration and docs
