+++ backend/piper_wrapper.py
@@ -84,7 +84,10 @@ class PiperWrapper:
         # create 1 second of silence as WAV
         with tempfile.NamedTemporaryFile(suffix=".wav") as tmp:
             tmp.write(
-                b"\x52\x49\x46\x46\x24\x00\x00\x00\x57\x41\x56\x45\x66\x6d\x74\x20\x10\x00\x00\x00\x01\x00\x01\x00\x44\xac\x00\x00\x88\x58\x01\x00\x02\x00\x10\x00\x64\x61\x74\x61\x00\x00\x00\x00"
+                b"\x52\x49\x46\x46\x24\x00\x00\x00\x57\x41\x56\x45"
+                b"\x66\x6d\x74\x20\x10\x00\x00\x00\x01\x00\x01\x00"
+                b"\x44\xac\x00\x00\x88\x58\x01\x00\x02\x00\x10\x00"
+                b"\x64\x61\x74\x61\x00\x00\x00\x00"
             )
             tmp.flush()
             audio = base64.b64encode(Path(tmp.name).read_bytes()).decode("ascii")
