+++ backend/tests/test_tts.py
@@ -2,9 +2,11 @@ import base64
 import os
 import sys
 
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
+sys.path.insert(
+    0, os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))
+)  # noqa: E402
 
-from app import app
+from app import app  # noqa: E402
 
 
 def test_tts_endpoint():
